pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aioworkers"
dynamic = [
    "version",
    "readme",
]
description = "Easy configurable workers based on asyncio"
license = "Apache-2.0"
requires-python = ">=3.8"
authors = [
    { name = "Alexander Malev", email = "malev@somedev.ru" },
]
keywords = [
    "aioworkers",
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Framework :: AsyncIO",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = []

[project.optional-dependencies]
aiohttp = [
    "aioworkers-aiohttp",
    "uvloop",
]
sentry = [
    "aioworkers-sentry",
]
prometheus = [
    "aioworkers-prometheus",
]
web = [
    "httptools",
    "uvloop",
    "pyyaml",
]
cron = [
    "crontab",
]
toml = [
    "tomli ; python_version < '3.11'",
]
interactive = [
    "ipykernel",
]
dev-test = [
    "coverage[toml]==7.6.1",
    "pytest==8.3.3",
    "pytest-aioworkers[asyncio,aiohttp]==0.4",
    "pytest-mock==3.14.0",
    "pytest-timeout==2.3.1",
    "pyyaml==6.0.2",
    "msgpack==1.1.0",
    "bson==0.5.10",
    "httptools==0.6.4",
]
dev-interactive = [
    "ipykernel==6.29.5",
    "ipython==8.30.0",
]
dev-lint = [
    "ruff==0.8.4",
    "mypy==1.13.0",
    "types-PyYAML",
    "types-setuptools",
]

[project.scripts]
aioworkers = "aioworkers.cli:main"

[project.urls]
Homepage = "https://github.com/aioworkers/aioworkers"
Documentation = "https://aioworkers.readthedocs.io"
Issues = "https://github.com/aioworkers/aioworkers/issues"
Source = "https://github.com/aioworkers/aioworkers"

[tool.hatch.version]
path = "aioworkers/__init__.py"

[tool.hatch.metadata.hooks.custom]
readme_content_type = "text/x-rst"
readme = "README.rst"
history = "HISTORY.rst"

[tool.hatch.build]
ignore-vcs = true
include = [
    "/aioworkers",
]

[tool.hatch.envs.default]
features = [
    "dev-test",
    "interactive",
    "cron",
]
[tool.hatch.envs.default.scripts]
cov = [
    "coverage run -m pytest",
    "coverage report",
]

[[tool.hatch.envs.all.matrix]]
python = [
    "3.8",
    "3.9",
    "3.10",
    "3.11",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = [
    "aioworkers",
    "tests",
]
addopts = [
    "-s",
    "--doctest-modules",
]

[tool.coverage.run]
branch = false
parallel = false
source = [
    "aioworkers",
    "tests",
]
omit = [
    "hatch_build.py",
]

[tool.coverage.report]
exclude_lines = [
    "no cov",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.ruff]
line-length = 120
indent-width = 4
target-version = "py38"

[tool.ruff.lint]
extend-select = [
  # pycodestyle
  "E",
  # Pyflakes
  "F",
  # pyupgrade
  # "UP",
  # flake8-bugbear
  "B",
  # flake8-simplify
  # "SIM",
  # isort
  "I",
]
extend-ignore = [
    "E501",
]

[tool.ruff.format]
quote-style = "preserve"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
follow_imports = "silent"
strict_optional = true
warn_redundant_casts = true
check_untyped_defs = true

[[tool.mypy.overrides]]
module = [
    "pytest",
    "aiohttp",
    "yarl",
    "setproctitle",
    "uvloop",
    "tomllib",
    "IPython.*",
    "ipykernel",
    "prompt_toolkit",
    "tornado",
]
ignore_missing_imports = true

[tool.hatch.envs.lint]
python = "3.8"
features = [
    "dev-lint",
]
[tool.hatch.envs.lint.scripts]
typing = "mypy --install-types --non-interactive {args:aioworkers tests}"
style = [
    "ruff check {args:aioworkers tests}",
    "ruff format --check --diff {args:aioworkers tests}",
]
fmt = [
    "ruff check --fix {args:aioworkers tests}",
    "ruff format {args:aioworkers tests}",
    "style",
]
all = [
    "style",
    "typing",
]

[tool.hatch.envs.template]
detached = true
dependencies = [
    "scaraplate",
]
[tool.hatch.envs.template.scripts]
update = [
    "scaraplate rollup --no-input {args:../cookiecutter-aioworkers-plugin} ."
]
clone = [
    "git clone https://github.com/aioworkers/cookiecutter-aioworkers-plugin {args:../cookiecutter-aioworkers-plugin}"
]
